+++ src/drvi/internal/drvi/__init__.py
@@ -1,0 +1,30 @@
+"""Developmental, internal-use-only DRVI extensions.
+
+.. warning::
+
+    Everything in :mod:`drvi.internal.drvi` is for **DRVI development only** and may change or be
+    removed without notice. For general use, import the maintained model as ``scvi.external.DRVI``
+    (or its alias ``drvi.model.DRVI``).
+
+:class:`drvi.internal.DRVI` subclasses :class:`scvi.external.DRVI` and re-adds a few experimental
+features kept from earlier drvi-py: opt-in residual connections between hidden layers, streaming
+(online) training metrics, and a sparse latent representation.
+"""
+
+from drvi.internal.drvi._base_components import DecoderDRVI, enable_residual
+from drvi.internal.drvi._generative_mixin import SparseLatentMixin
+from drvi.internal.drvi._metrics import LatentStats, StreamingPairwiseMI
+from drvi.internal.drvi._model import DRVI
+from drvi.internal.drvi._module import DRVIModule
+from drvi.internal.drvi._trainingplan import DRVITrainingPlan
+
+__all__ = [
+    "DRVI",
+    "DRVIModule",
+    "DecoderDRVI",
+    "DRVITrainingPlan",
+    "SparseLatentMixin",
+    "LatentStats",
+    "StreamingPairwiseMI",
+    "enable_residual",
+]
