+++ src/drvi/internal/drvi/_generative_mixin.py
